Crate bitter

Reading bits until the bitter end

Bitter reads bits in a desired endian format platform agnostically. Performance is the distinguishing feature compared to other bit readers. See benchmarks for more information.

Features

✔ support for reading and writing little endian, big endian, and native endian formats

✔ request an arbitrary amount of bits (up to 64 bits) and bytes

✔ ergonomic requests for common data types (eg: u8 … u64, f64, etc)

✔ fastest bit reader at multi-GiB/s throughput

✔ no allocations, no dependencies, and no panics

✔ no_std compatible

Quick start with Auto mode

The example below gives a good demonstration of the main API surface area to decode a 16 bit data model. It strikes a good balance between ergonomics and speed.

use bitter::{BitReader, LittleEndianReader};
let mut bits = LittleEndianReader::new(&[0xff, 0x04]);
assert_eq!(bits.bytes_remaining(), 2);
assert_eq!(bits.read_bit(), Some(true));
assert_eq!(bits.bytes_remaining(), 1);
assert_eq!(bits.read_u8(), Some(0x7f));
assert!(bits.has_bits_remaining(7));
assert_eq!(bits.read_bits(7), Some(0x02));

The read_ prefixed functions are colloquially known as “Auto mode”, as one does not need to manage the underlying bits in the lookahead buffer.

Manual mode

One can unlock additional performance by amortizing internal state logic management when exploiting patterns in the encoded data. Our example from before can be rewritten to take advantage of our domain knowledge that we’ll be decoding 16 bits.

use bitter::{BitReader, LittleEndianReader};
let mut bits = LittleEndianReader::new(&[0xff, 0x04]);

// ... snip code that may have read some bits

// We first check that there's enough total bits
if !bits.has_bits_remaining(16) {
  panic!("not enough bits remaining");
}

// Despite there being enough data, the lookahead buffer may not be sufficient
if bits.lookahead_bits() < 16 {
  bits.refill_lookahead();
  assert!(bits.lookahead_bits() >= 16)
}

// We use a combination of peek and consume instead of read_*
assert_eq!(bits.peek(1), 1);
bits.consume(1);

// Notice how the return type is not an `Option`
assert_eq!(bits.peek(8), 0x7f);
bits.consume(8);

// We can switch back to auto mode any time
assert!(bits.has_bits_remaining(7));
assert_eq!(bits.read_bits(7), Some(0x02));

The refill, peek, and consume combination are the building blocks for Manual Mode, and allows fine grain management for hot loops. The surface area of Manual Mode APIs is purposely compact to keep things simple. The Auto mode API is larger as that API should be the first choice.

A major difference between Manual mode and Auto mode is that one can’t peek at more than what’s in the lookahead buffer. Since the lookahead buffer will vary between MAX_READ_BITS and 63 bits, one will need to write logic to stitch together peeks above MAX_READ_BITS in the endian of their choice.

Manual mode can be intimidating, but it’s one of if not the fastest way to decode a bit stream, as it’s based on variant 4 from Fabian Giesen’s excellent series on reading bits. Others have employed this underlying technique to significantly speed up DEFLATE.

An example of how one can write a Manual mode solution for reading 60 bits:

use bitter::{BitReader, LittleEndianReader};

let data: [u8; 8] = [0xab, 0xcd, 0xef, 0x01, 0x23, 0x45, 0x67, 0x89];
let mut bits = LittleEndianReader::new(&data);

// ... snip ... maybe some bits are read here.

let expected = 0x0967_4523_01EF_CDAB;
let bits_to_read = 60u32;

bits.refill_lookahead();
let lo_len = bits.lookahead_bits();
let lo = bits.peek(lo_len);
bits.consume(lo_len);

let left = bits_to_read - lo_len;
bits.refill_lookahead();
let hi_len = bits.lookahead_bits().min(left);
let hi = bits.peek(hi_len);
bits.consume(hi_len);

assert_eq!(expected, (hi << lo_len) + lo);

The above is not an endorsement of the best way to simulate larger reads in Manual mode. For instance, it may be better to drain the lookahead first, or use MAX_READ_BITS to calculate lo instead of querying lookahead_bits. Always profile for your environment.

Unchecked mode

There’s one final trick that bitter exposes that dials performance to 11 at the cost of safety and increased assumptions. Welcome to the unchecked refill API (referred to as “unchecked”), which can only be called when there are at least 8 bytes left in the buffer. Anything less than that can cause invalid memory access. The upside is that this API unlocks the holy grail of branchless bit reading.

Always guard unchecked access at a higher level:

use bitter::{BitReader, LittleEndianReader, MAX_READ_BITS};

let mut bits = LittleEndianReader::new(&[0u8; 100]);
let objects_to_read = 10;
let object_bits = 56;
let desired_bits = objects_to_read * object_bits;
let bytes_needed = (desired_bits as f64 / 8.0).ceil();

if bits.unbuffered_bytes_remaining() >= bytes_needed as usize {
    for _ in 0..objects_to_read {
        unsafe { bits.refill_lookahead_unchecked() };
        let _field1 = bits.peek(2);
        bits.consume(2);

        let _field2 = bits.peek(18);
        bits.consume(18);

        let _field3 = bits.peek(18);
        bits.consume(18);

        let _field4 = bits.peek(18);
        bits.consume(18);
    }
} else if bits.has_bits_remaining(desired_bits) {
  // So have enough bits to read all the objects just not
  // enough bits to call the unchecked lookahead API everytime.
  assert!(false);
} else {
  // Not enough data.
  assert!(false);
}

All three modes: auto, manual, and unchecked can be mixed and matched as desired.

Write API

use bitter::{BitWriter, LittleEndianWriter};

let mut buffer = Vec::new();
{
    let mut writer = LittleEndianWriter::new(&mut buffer);

    // Write various data types
    writer.write_bit(true).unwrap();
    writer.write_u8(0xFF).unwrap();
    writer.write_u32(0x12345678).unwrap();
    writer.write_bits(5, 0x1F).unwrap();  // Write 5 bits
    writer.flush().unwrap();
}

println!("Wrote {} bytes", buffer.len());

The writer APIs act exactly like a BufWriter, as behind the scenes there is a bit buffer holding unflushed data, thus the same caveat with BufWriter applies here:

It is critical to call flush before [a writer] is dropped. Though dropping will attempt to flush the contents of the buffer, any errors that happen in the process of dropping will be ignored. Calling flush ensures that the buffer is empty and thus dropping will not even attempt file operations.

In the advent that a writer is flushed unaligned to a byte boundary, the remaining bits are padded with zeros.

no_std crates

This crate has a feature, std, that is enabled by default. To use this crate in a no_std context, add the following to your Cargo.toml:

[dependencies]
bitter = { version = "x", default-features = false }

Structs

BigEndianReader

Reads bits in the big-endian format

BigEndianWriter

Writes bits in the big-endian format

IntoInnerError

An error returned by calling into_inner on a writer.

LittleEndianReader

Reads bits in the little-endian format

LittleEndianWriter

Writes bits in the little-endian format

Remainder

Contains information about unread data remaining in a bit reader.

Constants

MAX_READ_BITS

The minimum number of bits available after a lookahead refill.

Traits

BitReader

Read bits in a given endian order

BitWriter

Write bits in a given endian order

Functions

sign_extend

Arbitrary sign extension for manual mode API.

Type Aliases

NativeEndianReader

Read bits in system native-endian format

NativeEndianWriter

Write bits in system native-endian format